Validate ranks/sizes match in HL sequencer copy op
In the HL we're still dealing with indices so it makes sense for the src and dst to have the same rank to ensure the right indices line up. This might mean we require creating redundant reshapes, but makes things more explicit without bleeding offsets through the index abstraction. After lowering to LL and offsets we should be able to fold away unnecessary reshaping.

PiperOrigin-RevId: 271999524