Add support for std.extract_element oepration in SPIR-V lowering.
Add limited support for std.extract_element operation in SPIR-V
lowering. Currently only supports handling extract element of
zero-rank tensors, but can be easily generalized to handle the general
case. The only unknown is to know more about how the operands of the
extract_element op look like at the time of SPIR-V lowering since they
are of index types, and XLA-HLO does not support index types. (The
lowering is fairly trivial and is left as a TODO for now).

PiperOrigin-RevId: 282504767