Build the compiler as part of the CMake build
This is actually necessary right now because otherwise we don't have tablegen

The alternative was https://github.com/google/iree/pull/307, which unconditionally included LLVM, but per https://github.com/google/iree/issues/357, we want to avoid requiring LLVM for the runtime. Since it is required right now, the simplest way to get things building is to just enable this as part of the action.

PiperOrigin-RevId: 287875037