Create unfoldable constant op.
This is just syntactic sugar around a std.constant wrapped in iree.do_not_optimize, but worthwhile sugar, I think, given that an unfoldable constant is our primary use case at the moment.

To ensure that translation still works, I needed to add canonicalization to the flow transform pipeline (and it has to be top-level, not a nested pass). I think if that's the approach we want to take, all the transform pipelines should probably include that, so I added it to all of them.

An alternative would be to put it in front of where those passes are stuck together (in each of the tools). I think it depends on the cost of canonicalization, whether the flow pipelines are expected to be user facing, and whether each pipeline is expected to *output* canonicalized IR.

PiperOrigin-RevId: 288136977