A GitHub action that prints out the author of a PR (#445)
I'm writing an action to auto-label PRs from the copybara app with "kokoro:run" so we don't have to keep doing it manually. I need to be able to filter for just the service because otherwise the action would try to label all PRs. For PRs coming from the main repo, this would just be redundant (everyone else with write access is a collaborator, so kokoro will run automatically) and may make kokoro confused. For PRs coming from forks it would just fail because they don't have the token to apply labels (GitHub's actions work terribly with forks).

So yeah testing in prod. I don't have any way to get the copybara app to send PRs to my fork. I could just write the action and hope, but I have this suspicion that I'll get it wrong. This action is pretty harmless.