Action to automatically update submodules (#437)
This will run after every push to master and synchronize the submodule state, using SUBMODULE_VERSIONS as the source of truth.

Due to constraints with the synchronization tooling from upstream, we cannot push a submodule update from upstream atomically. This will follow fast behind and synchronize the submodules.

There is the risk of races here with other things committing:
1. If this action doesn't finish pushing before another push occurs from elsewhere, it may fail (it will not force push). Since it will try again on the next push, this seems not too bad, and the price we must pay for using eventual consistency instead of atomic updates.
2. If this action completes a push and other automation is trying to push based on the previous HEAD, that automation may fail because the push would not be a fastfoward. In that case, we would also have to rely on the automation eventually retrying.

Let's try to avoid using this where possible (just update submodules as normal), but it does make upstream submodule updates possible, which seems nice.

Tested:
I tried this out in my fork using changes that did and did not update the SUBMODULE_VERSIONS file. You can see the history in this branch: https://github.com/GMNGeoffrey/iree/commits/submodule-autoupdate-demo (I actually did all this on master, but pulled it into a branch to preserve it). I'll leave the branch around for a while.