Make generated lit test target names match common internal format
It's generally good practice to prefix with the macro name, but internally bazel lit tests use this format and it's confusing for developers working across code bases to have both. We're unlikely to get conflicts here as we're probably not going to be generating multiple tests from the same source file.

PiperOrigin-RevId: 293231908