Modify XLA -> SPIR-V lowering to not rely on workload size.
The existing XLA to SPIR-V lowering relied on workload size to figure
out which threads within a workgroup is active. The result of this was
that there was an implicit assumption on the number of workgroups that
was built into the generated code. This was an implicit contract
between the code-generation and the runtime. Break this implicit
contract and modify the generated code to work with any number of
workgroups used by the runtime. The code is more branchy now, but that
can be canonicalized away later.

PiperOrigin-RevId: 296281206