Bazel to CMake: Mark linkopts unimplemented
These are not handled by the script, but should also be used in CMake.

Another option would be to error if there are any unrecognized kwargs and then handle all kwargs that we care about or are no-ops in CMake. The tradeoffs would depend on how many possible kwargs there are, how many we actually use, and how many are no-ops in CMake. Deferring that investigation. Even if we do this, we should probably accept the kwargs and then check their contents is empty so that we can choose whether to continue with a comment or error out depending on strict mode.

PiperOrigin-RevId: 298401856