shapex dialect constant folding hook
Any op that returns a fully defined ranked shape can be folded to a
shapex.const_ranked_shape op.

This generally cleans up IR and makes it easier to stare at.

Also, add OpAsmOpInterface for pretty names on the result of ConstRankedShapeOp. This makes testing a lot nicer, so that we can directly rely on e.g. `%rs3_4_i32` as the SSA value name for `!shapex.ranked_shape<[3,4],i32>`. It made fixing up the HLOToVMLA tests a lot easier (they were broken by this change since this changes canonicalization and that test runs `-canonicalize`).

PiperOrigin-RevId: 301953691