Add pass to tile and fuse linalg operations.
The current tiling pass is incorporated as a pattern. So if there is
nothing to fuse, the op is just tiled. The fallback of executing the
op sequentially is also incorprated as a pattern of this pass. The
pass added here also updates the workgroup size when the tiling
pattern succeeds. This makes the UpdateWorkGroupSizePass unnecessary.

PiperOrigin-RevId: 303810983