Add convenience wrappers for check ops allowing inline attribute
This makes the use site a bit less verbose. Implemented as a separate op with canonicalization to the existing ops to avoid an explosion of the API surface area for a single op. Future work could provide a single op with many options as the entry point and specific ops for the implementation that includes full verification.

Unfortunately, the current assembly format doesn't elide the attribute type because the type inference doesn't support AllTypesMatch with attributes. This will hopefully change soon and it can be dropped.

PiperOrigin-RevId: 305383915