Extract padding attributes from xla_hlo.reduce_window op.
We fold the pad op into a following reduce_window op because the pad op lowering
is not supported yet. Since the change has been landed, we can do it in a
reverse way -- extract the padding attributes out from the reduce_window op.
This can unblock the tiling/fusion issues in Linalg transforms because there
will not be any offset in the indexing maps.

PiperOrigin-RevId: 307938306