Add pass to establish invariant that each FuncOp is used <= 1 time.
This is an expedient workaround for the limitations of tf-shape-inference and gets us further on the ASR encoder model. The ASR encoder model triggers this in two places:
1. tf.While body func is reused in two places
2. The "do the LSTM layer" function is reused in two places.

It's sad that we need this, but as described in the code for the pass, there's actually two separate reasons we need it:
1. IREE requires a baseline level of shape inference to eliminate unranked, and currently tf-shape-inference is the only place where we can plausible do that.
2. Cloning / specializing is actually a pretty valid thing to do just for general IR health in many cases, such as the "do the LSTM layer" functions which have different hyperparameters. In fact, most of the "call" places that this pass will clone are in practice going to end up inlined later.

I initially thought that this would all be fixed by just upgrading tf-shape-inference to support CFG (and started looking at doing that), but in fact it makes a domain-specific assumption about tf.While, namely that the types of the loop carried values do not change. The tf.While op verifier specifically allows the more general behavior, but many ways to create such loops in practice do conform (which is probably why this pass seems to work in practice). Autographed loops conform, and tf.while_loop's that have not been explicitly created with shape_invariants kwarg conform (see https://www.tensorflow.org/api_docs/python/tf/while_loop). Even if we did apply this assumption when lowering to CFG, a naive extension of tf-shape-inference would not easily do this inference, because it uses what is called a "pessimistic" algorithm which cannot refine the types of use-def cycles (such as loops), and would have to be changed to an "optimistic" algorithm which would be a lot of work. We could do something hacky there like assume that all types flowing along backedges will eventually become consistent with the forward edges. For now, this hack seems slightly preferable.

PiperOrigin-RevId: 309847068