[codegen][llvm] Collect iree.placeholder from all blocks
When converting from Linalg on tensors to Linalg on buffers, we create
iree.placeholder op just before the Linalg op using it. It means if we
have multiple Linalg ops it will end up we have iree.placeholder ops
scattered in the middle of the IR. When going to the LLVM CodeGen, we
insert branches and have multiple blocks so that means some iree.placeholder
ops are not in the entry block anymore. The ProcessFuncInterfacePattern
was only collecting iree.placeholder ops from the entry block. It ends up
after the pattern is applied we missed some iree.placeholder ops and that
results in the generated FuncOp to still be illegal.

Fixes https://github.com/google/iree/issues/1838

PiperOrigin-RevId: 310464616