Resubmitting PR #2163 with fixes.
This commit is a squash of two commits. The first is the same change
PR #2163.

The second commit contains fixes that address failures on Resnet seen
from that P. The original commit seems to have a correctness
issue. The change is valid if the number of iterations of the loop is
less than or equal to the workgroup size, which doesnt seem to be the
case for convolution/pooling in all cases. More investigation is
needed. For now, falling back to the loop method for
convolution/pooling when there is no padding. With padding, tiling is
completely avoided and is executed in parallel by linearizing all the
parallel loops and distributing to threads using the global invocation
ID. The commit also enables the option to split the padding into a
separate option when needed. This is done to make it easier to
pin-point the issue, so that the complications from padding are
removed.

PiperOrigin-RevId: 318839822