Avoid recompiling e2e modules with global variables (#2897)
Follow-up to #2899. The global variables `_global_ref_module` and `_global_tar_modules` are used to store compiled IREE modules for the `TracedModuleTestCase`. Storing the variables on the `TracedModuleTestCase` `cls` causes the errors in #2900, and storing the variables on the `TracedModuleTestCase` instance (via `self`) doesn't work because `tf.test.TestCase` wipes these variables before each unittest.