Add a pass to vectorize Linalg ops. (#3107)
The convert op is the only one not covered, so we have an extra flag to
enable the pass. It needs an upstream fix in MLIR.

A Linalg generic op will be vectorized only if:
  1) All the memory access is contiguous.
  2) The shape of the last dim is a multiple of 4.
  3) All the operands have only one use, which is the generic op.

This pass can also be extended to create large vectors. In the patch, it
only consider small vectors like 4xf32/4xi32.