Working around lack of mlir symbol side effects for variable loads. (#3498)
This turns flow/hal/vm variable loads into pure ops when the variable they
reference is immutable. This allows CSE and DCE to work its magic.

Partially helps with #1145, though only in specific cases of CSE-able
values. Thankfully the mix of CSE-able to non-CSE-able is about 1000:1 in
BERT, so this is worth doing (and can be easily replaced with symbol-based
side-effects when MLIR gains them).