Add script to symbolize android stacks, particularly for use with Asan. (#3673)
* Add script to symbolize android stacks, particularly for use with Asan.

You just paste some ASan output to a text file and feed that to stdin to
this script. It will then pull the files out of the device and symbolize
them on the host using the NDK-provided host build of llvm-symbolizer.
Because it just pulls files from the device, there's no need to point
this script to any particular obj dir. It's easy to use.

ANDROID_NDK=~/android-ndk-r21d ~/android-symbolize.sh < ~/a

Discussion of alternatives: https://github.com/android/ndk/issues/753

* rename