Add an "executable tutorial" going from zero to tracy on android (#3664)
* Add an "executable tutorial" going from zero to tracy on android
benchmark.

The idea is that a beginner would run it a few times, then it'd be
mostly useful as a tutorial to read steps from, not intended to be used
long term as a black box tool.

Just try running it. It'll complain a bit that you need to define
environment variables IREE_ROOT and ANDROID_NDK, and once you do that,
it should figure all the rest, including building tracy UI, going
through the whole steps of compiling an android benchmark workload,
running that in tracy, setting up port forwarding, and finally telling
you to run a command to run your (newly built) tracy UI to connect
and show the trace.