**KWS Streaming Changes**
Add function for creating a model with default params.

Including this function in the source code also documents the most basic way to create a new model and changing it's streaming mode, which can serve as a jumping off point for new users.

**IREE Changes**

Update `BUILD.overlay` and import for affected test internally.

PiperOrigin-RevId: 340526468