Update Vulkan SDK within CI docker images (#3825)
This commit changes to download the Vulkan SDK tarball instead of
installing it from APT repo so that we can still get it for the
RBE toolchain. LunarG APT repo dropped support for Ubuntu 16.04
for recent SDK versions.

This commit also creates a base Vulkan SDK image so that other
docker images can copy the SDK from it.

This commit also pin the NVIDIA driver and `tf-nightly` to a specific
version to increase stability. A previous pull request
https://github.com/google/iree/pull/3754 with mostly the same
change was reverted due to failures on TensorFlow integration
tests. That should be fixed now by pinning down the `tf-nightly`
version.