Remove apt repository for wrong Ubuntu version (#4302)
"eoan" is Ubuntu 19.04 and these containers are Ubuntu 18.04. This is
causing failures in the action complaining that "The repository does
not have a Release file." (e.g. 
https://github.com/google/iree/pull/4301/checks?check_run_id=1601595431)
We could replace with "bionic", but this
actually doesn't seem to be necessary at all. It's been in place since
https://github.com/google/iree/commit/d8b66f2b7b1c and
https://github.com/google/iree/commit/33063e4224fc when we were running
with the "latest" ubuntu containers, which probably explains why they
were pointing at a different version.