Support for TF.Gather including intermediate dialect work (#4292)
Gather operations added to the strings and tf_strings dialect with
lowerings from TensorFlow to tf_strings and tf_string to strings.
This includes tests for each and an end-to-end test. Right now the
python test does not validate values (missing conversion from
StringTensor to Numpy equivalent) but prints instead. Close enough
for current support.