Avoid running slow benchmarks as part of testing (#4364)
It's good to test that these benchmarks run. It's really annoying to run
a billion iterations of them every time you try to run tests. This
literally takes longer than all the other tests combined. Instead, make
these binaries and invoke them for testing with
`--min_benchmark_time=0`.

This introduces CMake and Bazel -> CMake support for `run_binary_test`,
which runs the specified executable target with the specified args.
Punted on support for data dependencies, since these are not required
for this use case and I think we still haven't figured out how to do
that for CMake tests.