Force CPU benchmark to run on single thread. (#4751)
In the past, we only could run CPU on single thread. After 1a56cad
landed, we are able to run on multiple threads, and it's turned on by
default. The PR forces it to run on single thread to keep the same
behavior.

Fixes https://github.com/google/iree/issues/4686