Action to fixup Copybara failed merges (#4744)
Extend the submodule synchronization action to fix various issues with
Copybara-exported PRs. It rewrites commits with the Copybara tag to
mark them as merged from the specified commit.

When this is rolled out, any commit pushed that contains the Copybara
tag COPYBARA_INTEGRATE_REVIEW, which indicates that Copybara did not
successfully perform an integrate (because it didn't remove this tag)
will be rewritten to merge from the specified commit and update the
submodules from SUBMODULE_VERSIONS.txt.

Tested:
Ran this on my fork. Pushed
https://github.com/GMNGeoffrey/iree/commit/81d31db7b8f4
![fake_copybara_push](https://user-images.githubusercontent.com/5732088/106990286-ce27ce00-6728-11eb-9e6f-def4e644d963.png)
![fake_copybara_commit](https://user-images.githubusercontent.com/5732088/106990453-24950c80-6729-11eb-9f2c-805225a45142.png)
which triggered
https://github.com/GMNGeoffrey/iree/runs/1836274764 which created
https://github.com/GMNGeoffrey/iree/commit/462987384b70
![rewritten_fake_copybara_push](https://user-images.githubusercontent.com/5732088/106990668-8e151b00-6729-11eb-9390-220b60b14a91.png)
![rewritten_fake_copybara_commit](https://user-images.githubusercontent.com/5732088/106990686-99684680-6729-11eb-9b05-aa9d13a6673d.png)