Replace all globs in IREE core with `enforce_glob` (#5063)
This avoids any globs of source files in CMake (which are
[discouraged](https://cmake.org/cmake/help/latest/command/file.html#glob))
and any globs being evaluated in bazel_to_cmake (which would make it
depend on files other than the BUILD file). It still allows the safety
check that you actually included all the files you meant to (which is
particularly useful with tests, where a failure to do so is test
skipped forever instead of an immediate build failure). The cost is
having to explicitly list a new source file when you add it, which
seems not so bad.

I didn't remove bazel_to_cmake support for glob yet, so I can clean up
any stragglers after this lands.

Fixes https://github.com/google/iree/issues/1083