(Re-)Open source IREE e2e BUILD configs
Note that these are *not buildable with Bazel*. We are still working on
a long-term plan for integration testing. These are open-sourced only
to enable Bazel->CMake conversion to restore coverage for our
integration tests in open source. As such, they include deps and other
things that only make sense in google's internal source tree.

PiperOrigin-RevId: 363026562